+++ src/managers/InfoManager.ts
@@ -2,10 +2,11 @@ import fetch from "isomorphic-fetch"
 import { ManagerOptions, WrapperOptions } from "../types/types"
 import { Webhook } from "fagc-api-types"
 import BaseManager from "./BaseManager"
-import { GenericAPIError, NoAuthError } from "../types"
+import { GenericAPIError } from "../types"
 import strictUriEncode from "strict-uri-encode"
 import { APIEmbed } from "discord-api-types"
 import { FetchRequestTypes } from "../types/privatetypes"
+import { MasterAuthenticate } from "../utils"
 
 export default class InfoManager extends BaseManager<Webhook> {
 	constructor(options: WrapperOptions, managerOptions: ManagerOptions = {}) {
@@ -50,6 +51,7 @@ export default class InfoManager extends BaseManager<Webhook> {
 		return add
 	}
 
+	@MasterAuthenticate()
 	async notifyGuildText({
 		guildId,
 		text,
@@ -58,13 +60,6 @@ export default class InfoManager extends BaseManager<Webhook> {
 		guildId: string,
 		text: string,
 	} & FetchRequestTypes): Promise<void> {
-		if (
-			!this.masterapikey &&
-			!reqConfig.masterapikey &&
-			!reqConfig.cookieAuth
-		)
-			throw new NoAuthError()
-
 		await fetch(
 			`${this.apiurl}/informatics/notify/${strictUriEncode(guildId)}`,
 			{
@@ -74,14 +69,13 @@ export default class InfoManager extends BaseManager<Webhook> {
 				}),
 				credentials: "include",
 				headers: {
-					authorization: !reqConfig.cookieAuth
-						? `Token ${reqConfig.masterapikey || this.masterapikey}`
-						: "Cookie",
+					authorization: `${reqConfig._keystring}`,
 					"content-type": "application/json",
 				},
 			}
 		)
 	}
+	@MasterAuthenticate()
 	async notifyGuildEmbed({
 		guildId,
 		embed,
@@ -90,13 +84,6 @@ export default class InfoManager extends BaseManager<Webhook> {
 		guildId: string,
 		embed: APIEmbed,
 	} & FetchRequestTypes): Promise<void> {
-		if (
-			!this.masterapikey &&
-			!reqConfig.masterapikey &&
-			!reqConfig.cookieAuth
-		)
-			throw new NoAuthError()
-
 		await fetch(
 			`${this.apiurl}/informatics/notify/${strictUriEncode(
 				guildId
@@ -106,9 +93,7 @@ export default class InfoManager extends BaseManager<Webhook> {
 				body: JSON.stringify(embed),
 				credentials: "include",
 				headers: {
-					authorization: !reqConfig.cookieAuth
-						? `Token ${reqConfig.masterapikey || this.masterapikey}`
-						: "Cookie",
+					authorization: `${reqConfig._keystring}`,
 					"content-type": "application/json",
 				},
 			}