+++ changelog.md
@@ -1,3 +1,9 @@
+## UNRELEASED
+
+### Changes
+
+-	Remove profile and user managers
+
 ## [1.23.0] - [2022-01-03]
 
 ### Changes