+++ changelog.md
@@ -1,9 +1,10 @@
-## UNRELEASED
+## [1.24.0] - [2022-01-03]
 
 ### Changes
 
 -	Remove profile and user managers
 -	Use decorators for handling of authentication in one place
+-	Update managers to new API
 
 ## [1.23.0] - [2022-01-03]